# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrainCoach API",
    "version": "v1"
  },
  "paths": {
...
    "/api/athletes/{athleteUserId}/plans": {
      "get": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingPlanDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingPlanDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingPlanDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{id}": {
      "get": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPlanDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPlanDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPlanDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans": {
      "post": {
        "tags": [
          "TrainingPlans"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingPlanRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingPlanRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingPlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPlanDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPlanDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingPlanDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{planId}/weeks": {
      "post": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingWeekRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingWeekRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingWeekRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingWeekDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingWeekDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingWeekDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/weeks/{weekId}": {
      "put": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "weekId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrainingWeekRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrainingWeekRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrainingWeekRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingWeekDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingWeekDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingWeekDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/{id}": {
      "get": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedWorkoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedWorkoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlannedWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/workouts": {
      "post": {
        "tags": [
          "TrainingPlans"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedWorkoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedWorkoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlannedWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/{id}/copy": {
      "post": {
        "tags": [
          "TrainingPlans"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyWorkoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyWorkoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CopyWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannedWorkoutDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CopyWorkoutRequest": {
        "type": "object",
        "properties": {
          "targetTrainingWeekId": {
            "type": "string",
            "format": "uuid"
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "CreatePlannedWorkoutRequest": {
        "type": "object",
        "properties": {
          "trainingWeekId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "sport": {
            "$ref": "#/components/schemas/SportType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "coachDescription": {
            "type": "string",
            "nullable": true
          },
          "isRestDay": {
            "type": "boolean"
          },
          "plannedDistanceMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "plannedDurationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "plannedElevationGainMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutSegmentDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateTrainingPlanRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "seasonId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateTrainingWeekRequest": {
        "type": "object",
        "properties": {
          "weekStartDate": {
            "type": "string",
            "format": "date"
          },
          "weekIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "IntensityTargetType": {
        "enum": [
          "Free",
          "HeartRateZone",
          "Pace",
          "Rpe",
          "Power"
        ],
        "type": "string"
      },
...
      "PlannedWorkoutDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "trainingWeekId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "sport": {
            "$ref": "#/components/schemas/SportType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "coachDescription": {
            "type": "string",
            "nullable": true
          },
          "isRestDay": {
            "type": "boolean"
          },
          "plannedDistanceMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "plannedDurationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "plannedElevationGainMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutSegmentDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SportType": {
        "enum": [
          "Running",
          "Cycling",
          "Swimming",
          "Strength",
          "CrossTraining",
          "Rest",
          "Other"
        ],
        "type": "string"
      },
...
      "TrainingPlanDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "coachUserId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "weeks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingWeekDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TrainingPlanDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "coachUserId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "TrainingWeekDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "trainingPlanId": {
            "type": "string",
            "format": "uuid"
          },
          "weekStartDate": {
            "type": "string",
            "format": "date"
          },
          "weekIndex": {
            "type": "integer",
            "format": "int32"
          },
          "coachWeekSummary": {
            "type": "string",
            "nullable": true
          },
          "athleteWeekReflection": {
            "type": "string",
            "nullable": true
          },
          "whatWasMissingNote": {
            "type": "string",
            "nullable": true
          },
          "athleteWeeklyRating": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "workouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlannedWorkoutDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePlannedWorkoutRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "sport": {
            "$ref": "#/components/schemas/SportType"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "coachDescription": {
            "type": "string",
            "nullable": true
          },
          "isRestDay": {
            "type": "boolean"
          },
          "plannedDistanceMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "plannedDurationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "plannedElevationGainMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutSegmentDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTrainingWeekRequest": {
        "type": "object",
        "properties": {
          "coachWeekSummary": {
            "type": "string",
            "nullable": true
          },
          "athleteWeekReflection": {
            "type": "string",
            "nullable": true
          },
          "whatWasMissingNote": {
            "type": "string",
            "nullable": true
          },
          "athleteWeeklyRating": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WorkoutSegmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/WorkoutSegmentType"
          },
          "repeatCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "distanceMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "intensityTargetType": {
            "$ref": "#/components/schemas/IntensityTargetType"
          },
          "targetHeartRateZoneId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "targetPaceSecondsPerKmMin": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetPaceSecondsPerKmMax": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetRpe": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetPowerWatts": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkoutSegmentType": {
        "enum": [
          "WarmUp",
          "Main",
          "Interval",
          "Rest",
          "CoolDown",
          "Repeat",
          "Drill",
          "Strides"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Zadejte JWT access token (bez prefixu \"Bearer \").",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}